requirements.txt
# Machine Learning & Vision
torch
transformers==4.36.2
pillow==10.2.0
numpy==1.24.3

# Vector Database
qdrant-client==1.13.2

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
lxml==5.1.0

# Database
pymongo==4.6.1

# API
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9

# Utils
python-dotenv==1.0.0

# Prerequisites:
#
# 2. Install Ollama from https://ollama.ai/download
#    ollama pull llava
#    ollama serve
#    ollama run llava
#
# 3. Set up Qdrant (local or cloud)
#    Local: docker run -p 6333:6333 qdrant/qdrant
#    Cloud: Set QDRANT_URL and QDRANT_API_KEY in .env
#
# 4. Run MongoDB:
#    mongod

# Environment variables needed in .env:
# MONGODB_URI=mongodb://localhost:27017
# QDRANT_URL=http://localhost:6333
# QDRANT_API_KEY=your_api_key  # Only for cloud deployment
